created an api route for the courses
wrote a quick api route that returns information about all the courses in the database (in the test database in this commit, the only course is algebra I) in JSON format. will now port the courses page to django template format so that josh can work on that page as well as the home page.