converted the courses.html to django template format
i converted the courses.html to the django template format and fixed the asset connection. next is to complete the api route and dynamically populate the courses there. also, while doing this, i realized we're also going to need to do something similar api route wise with the categories so the categories aren't hard coded but dynamic. i will work on that next.